+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 *   Moved database queries to a web worker to gain more responsive UI
 *   Fixed Gelbooru, Instagram, and Twitter previews
 *   Fixed green names not showing up when 'show friends/bookmarks in a different colour' is selected
+*   Sped up Imgur previews
+*   Minor UI design adjustments for user names and badges
 
 
 ## 1.9.0